test: split shell contract suite
